Working on client and server now! Added in bark mechanic for making latex early game. Rubber tree shape is better.

+++ src/main/java/theking530/staticpower/cables/attachments/digistore/exporter/DigistoreExporterAttachment.java
@@ -185,7 +185,7 @@ public class DigistoreExporterAttachment extends AbstractCableAttachment {
 	protected int getExportRate(ItemStack attachment) {
 		float acceleratorCardCount = getUpgradeCount(attachment, AcceleratorUpgrade.class);
 		if (acceleratorCardCount > 0) {
-			float accelerationAmount = StaticPowerConfig.acceleratorCardMaxImprovment * (acceleratorCardCount / ModUpgrades.AcceleratorUpgrade.getMaxStackSize());
+			double accelerationAmount = StaticPowerConfig.acceleratorCardMaxImprovment * (acceleratorCardCount / ModUpgrades.AcceleratorUpgrade.getMaxStackSize());
 			return (int) (StaticPowerConfig.digistoreExporterRate / accelerationAmount);
 		} else {
 			return StaticPowerConfig.digistoreExporterRate;
